+++ src/main/java/com/blog/myblog/domain/ConcurrencySafeEntity.java
@@ -4,9 +4,9 @@ import java.util.ConcurrentModificationException;
 
 public interface ConcurrencySafeEntity {
 
-    Short version();
+    Integer version();
 
-    default void detectConcurrentModification(Short transmittedVersion){
+    default void detectConcurrentModification(Integer transmittedVersion){
        if(!version().equals(transmittedVersion)){
            throw new ConcurrentModificationException("This entity was modified by another request in the meantime");
        }
